requirements.txt
streamlit
transformers
torch-cpu
pillow